@@ -19,8 +19,8 @@
   },
   "devDependencies": {
     "@types/node": "^11.9.5",
-    "tslint": "^5.13.0",
-    "typescript": "^3.3.3333"
+    "tslint": "^5.19.0",
+    "typescript": "^3.6.2"
   },
   "dependencies": {
     "sqrl": "file:../../packages/sqrl"
