@@ -8,21 +8,20 @@
     "registry": "https://registry.npmjs.org"
   },
   "dependencies": {
-    "lerna": "^3.10.5"
+    "lerna": "^3.15.0"
   },
   "devDependencies": {
-    "@types/bignum": "^0.0.29",
-    "@types/ioredis": "^4.0.4",
-    "@types/jest": "^24.0.9",
-    "@types/jest-diff": "^20.0.0",
+    "@types/ioredis": "^4.0.13",
+    "@types/jest": "^24.0.15",
+    "@types/jest-diff": "^20.0.1",
     "@types/jsonpath": "^0.2.0",
-    "@types/node": "^11.9.5",
+    "@types/node": "^12.6.3",
     "@types/pegjs": "^0.10.1",
-    "@types/prettier": "^1.15.2",
-    "coveralls": "^3.0.2",
-    "jest": "^24.3.1",
+    "@types/prettier": "^1.16.4",
+    "coveralls": "^3.0.5",
+    "jest": "^24.8.0",
     "jest-environment-node-debug": "^2.0.0",
-    "jest-extended": "^0.11.0",
+    "jest-extended": "^0.11.2",
     "pegjs": "^0.10.0",
     "rimraf": "^2.6.3",
     "sqrl": "file:packages/sqrl",
@@ -37,13 +36,13 @@
     "sqrl-test-utils": "file:packages/sqrl-test-utils",
     "sqrl-text-functions": "file:packages/sqrl-text-functions",
     "sqrl-website": "file:website",
-    "ts-jest": "^24.0.0",
-    "ts-node": "^8.0.2",
-    "ts-pegjs": "^0.2.2",
-    "tslint": "^5.12.1",
-    "tslint-config-prettier": "^1.17.0",
-    "typedoc": "git+https://github.com/TypeStrong/typedoc.git#7424d9705dae528e4ec111a176e3c3d91364f8e5",
-    "typescript": "^3.2.4",
+    "ts-jest": "^24.0.2",
+    "ts-node": "^8.3.0",
+    "ts-pegjs": "^0.2.5",
+    "tslint": "^5.18.0",
+    "tslint-config-prettier": "^1.18.0",
+    "typedoc": "^0.15.0",
+    "typescript": "^3.5.3",
     "wikipedia-diff-stream": "file:packages/wikipedia-diff-stream"
   },
   "scripts": {
