@@ -41,32 +41,31 @@
     "watch:test:integration": "npm run test:integration -- --watch"
   },
   "devDependencies": {
-    "@types/jest": "^24.0.9",
+    "@types/jest": "^24.0.18",
     "@types/node": "^11.9.5",
-    "jest": "^24.3.1",
+    "jest": "^24.9.0",
     "request-promise-native": "^1.0.7",
-    "strip-ansi": "^5.0.0",
-    "ts-jest": "^24.0.0",
-    "ts-node": "^8.0.2",
-    "typescript": "^3.2.4"
+    "strip-ansi": "^5.2.0",
+    "ts-jest": "^24.0.2",
+    "ts-node": "^8.3.0",
+    "typescript": "^3.6.2"
   },
   "engines": {
     "node": ">=6.0.0"
   },
   "dependencies": {
-    "@types/chokidar": "^1.7.5",
     "chalk": "^2.4.2",
-    "chokidar": "^2.0.4",
-    "csv-stringify": "^5.2.0",
+    "chokidar": "^3.0.2",
+    "csv-stringify": "^5.3.3",
     "docopt": "^0.6.2",
-    "eventemitter3": "^3.1.0",
+    "eventemitter3": "^4.0.0",
     "expand-tilde": "^2.0.2",
     "fast-stable-stringify": "^1.0.0",
-    "micro": "^9.3.3",
+    "micro": "^9.3.4",
     "micro-query": "^0.3.0",
     "micro-route": "^2.5.0",
-    "moment": "^2.23.0",
-    "split2": "^3.1.0",
+    "moment": "^2.24.0",
+    "split2": "^3.1.1",
     "sqrl": "file:../sqrl",
     "sqrl-cli-functions": "file:../sqrl-cli-functions",
     "sqrl-common": "file:../sqrl-common",
