@@ -37,20 +37,20 @@
     "watch:test:integration": "npm run test:integration -- --watch"
   },
   "devDependencies": {
-    "@types/jest": "^24.0.9",
+    "@types/jest": "^24.0.18",
     "@types/node": "^11.9.5",
-    "jest": "^24.3.1",
+    "jest": "^24.9.0",
     "sqrl-load-functions": "file:../sqrl-load-functions",
-    "ts-jest": "^24.0.0",
-    "ts-node": "^8.0.2",
-    "typescript": "^3.2.4"
+    "ts-jest": "^24.0.2",
+    "ts-node": "^8.3.0",
+    "typescript": "^3.6.2"
   },
   "engines": {
     "node": ">=6.0.0"
   },
   "dependencies": {
     "node-nice": "0.0.1",
-    "re2": "^1.8.4",
+    "re2": "^1.10.0",
     "simhash-js": "^1.0.0",
     "sqrl": "file:../sqrl"
   }
