@@ -3,11 +3,8 @@
  * Licensed under the Apache License, Version 2.0
  * http://www.apache.org/licenses/LICENSE-2.0
  */
-import bignum = require("bignum");
+import { toBufferBE } from "bigint-buffer";
 
 export function timeToBuffer(timeMs: number) {
-  return new bignum(timeMs).toBuffer({
-    endian: "big",
-    size: 8
-  });
+  return toBufferBE(BigInt(timeMs), 8);
 }
